refactor final tests, fixes #43
