+++ test/TypeRight.Tests/TestsWithoutParsing/MiscHelperTests.cs
